refactor: move boilerplate to libraries
